refactor(payload): extract shared walkSchema helper consumed by buildComponentIndex and buildImportMaps
