+++ packages/payload/src/admin/buildImportMaps.ts
@@ -1,6 +1,9 @@
 import type { ComponentSlot } from '../config/buildComponentIndex.js'
 import type { PayloadComponent, SanitizedConfig } from '../config/types.js'
-import type { Field, Tab } from '../fields/config/types.js'
+import type { Field } from '../fields/config/types.js'
+
+import { COMPONENT_SLOTS } from '../config/buildComponentIndex.js'
+import { walkSchema } from '../config/walkSchema.js'
 
 export type ImportMapEntryKind = 'admin-condition' | 'admin-validate' | 'component' | 'validate'
 
@@ -19,127 +22,17 @@ export type ImportMaps = {
 
 const INLINE_PATH_MARKER = '<inline>'
 
-const COMPONENT_SLOTS: ComponentSlot[] = [
-  'Field',
-  'beforeInput',
-  'afterInput',
-  'RowLabel',
-  'Label',
-  'Description',
-  'Error',
-]
-
 export function buildImportMaps(config: SanitizedConfig): ImportMaps {
   const serverEntries: ImportMapEntry[] = []
   const clientEntries: ImportMapEntry[] = []
 
-  for (const collection of config.collections ?? []) {
-    walkFields({
-      clientEntries,
-      fields: collection.fields ?? [],
-      pathSegments: [collection.slug],
-      serverEntries,
-    })
-  }
-
-  for (const global of config.globals ?? []) {
-    walkFields({
-      clientEntries,
-      fields: global.fields ?? [],
-      pathSegments: [global.slug],
-      serverEntries,
-    })
-  }
+  walkSchema(config, ({ field, fieldPath }) => {
+    collectFieldEntries({ clientEntries, field, fieldPath, serverEntries })
+  })
 
   return {
     client: { entries: clientEntries },
     server: { entries: serverEntries },
-  }
-}
-
-// Schema walker for two-import-map emission. Keep in sync with the
-// equivalent walker in packages/payload/src/config/buildComponentIndex.ts
-// until both are extracted into a shared helper.
-function walkFields({
-  clientEntries,
-  fields,
-  pathSegments,
-  serverEntries,
-}: {
-  clientEntries: ImportMapEntry[]
-  fields: Field[]
-  pathSegments: string[]
-  serverEntries: ImportMapEntry[]
-}): void {
-  for (const field of fields) {
-    const fieldPathSegments = computeFieldPath(field, pathSegments)
-    const fieldPath = fieldPathSegments.join('.')
-
-    collectFieldEntries({
-      clientEntries,
-      field,
-      fieldPath,
-      serverEntries,
-    })
-
-    switch (field.type) {
-      case 'array': {
-        walkFields({
-          clientEntries,
-          fields: field.fields ?? [],
-          pathSegments: [...fieldPathSegments, '*'],
-          serverEntries,
-        })
-        break
-      }
-      case 'blocks': {
-        const blocks = (field.blocks ?? []).filter(
-          (block): block is Exclude<typeof block, string> => typeof block !== 'string',
-        )
-        for (const block of blocks) {
-          walkFields({
-            clientEntries,
-            fields: block.fields ?? [],
-            pathSegments: [...fieldPathSegments, '*'],
-            serverEntries,
-          })
-        }
-        break
-      }
-      case 'collapsible':
-      case 'row': {
-        walkFields({
-          clientEntries,
-          fields: field.fields ?? [],
-          pathSegments,
-          serverEntries,
-        })
-        break
-      }
-      case 'group': {
-        walkFields({
-          clientEntries,
-          fields: field.fields ?? [],
-          pathSegments: fieldPathSegments,
-          serverEntries,
-        })
-        break
-      }
-      case 'tabs': {
-        for (const tab of field.tabs ?? []) {
-          const tabSegments = isNamedTab(tab) ? [...pathSegments, tab.name] : pathSegments
-          walkFields({
-            clientEntries,
-            fields: tab.fields ?? [],
-            pathSegments: tabSegments,
-            serverEntries,
-          })
-        }
-        break
-      }
-      default:
-        break
-    }
   }
 }
 
@@ -249,17 +142,6 @@ function resolveComponentPath(component: PayloadComponent | undefined): string |
     return component.path
   }
   return undefined
-}
-
-function computeFieldPath(field: Field, parentSegments: string[]): string[] {
-  if ('name' in field && typeof field.name === 'string' && field.name.length > 0) {
-    return [...parentSegments, field.name]
-  }
-  return parentSegments
-}
-
-function isNamedTab(tab: Tab): tab is Extract<Tab, { name: string }> {
-  return typeof (tab as { name?: unknown }).name === 'string'
 }
 
 function normalizeAdminValidateRef(ref: unknown): { exportName?: string; path: string } | null {
